Added Distributed Locking

+++ src/He.PipelineAssessment.UI/Features/SinglePipeline/Sync/SyncCommandHandler.cs
@@ -7,14 +7,16 @@ namespace He.PipelineAssessment.UI.Features.SinglePipeline.Sync
 {
     public class SyncCommandHandler : IRequestHandler<SyncCommand, SyncModel>
     {
+        private readonly IDistributedLockRepository _distributedLockRepository;
         private readonly IAssessmentRepository _assessmentRepository;
         private readonly ISinglePipelineProvider _singlePipelineProvider;
         private readonly ISyncCommandHandlerHelper _syncCommandHandlerHelper;
         private readonly ILogger<SyncCommandHandler> _logger;
 
-        public SyncCommandHandler(IAssessmentRepository assessmentRepository, ISinglePipelineProvider singlePipelineProvider, ISyncCommandHandlerHelper syncCommandHandlerHelper, ILogger<SyncCommandHandler> logger)
+        public SyncCommandHandler(IAssessmentRepository assessmentRepository , IDistributedLockRepository distributedLockRepository, ISinglePipelineProvider singlePipelineProvider, ISyncCommandHandlerHelper syncCommandHandlerHelper, ILogger<SyncCommandHandler> logger)
         {
             _assessmentRepository = assessmentRepository;
+            _distributedLockRepository = distributedLockRepository;
             _singlePipelineProvider = singlePipelineProvider;
             _syncCommandHandlerHelper = syncCommandHandlerHelper;
             _logger = logger;
@@ -23,45 +25,57 @@ namespace He.PipelineAssessment.UI.Features.SinglePipeline.Sync
         public async Task<SyncModel> Handle(SyncCommand request, CancellationToken cancellationToken)
         {
             var syncModel = new SyncModel();
-            try
+            if(await _distributedLockRepository.TryAcquireLockAsync("bg_task"))
             {
-                var data = await _singlePipelineProvider.GetSinglePipelineData();
-                if (data.Any())
+                try
                 {
-                    List<int> sourceAssessmentSpIds = data.Select(x => x.sp_id!.Value).ToList();
+                    var data = await _singlePipelineProvider.GetSinglePipelineData();
+                    if (data.Any())
+                    {
+                        List<int> sourceAssessmentSpIds = data.Select(x => x.sp_id!.Value).ToList();
 
-                    var destinationAssessments = await _assessmentRepository.GetAssessments();
-                    var destinationAssessmentSpIdsThatHaveBeganAssessment = destinationAssessments.Where(x => x.AssessmentToolWorkflowInstances != null && x.AssessmentToolWorkflowInstances.Any())
-                        .Select(x => x.SpId).ToList();
-                    var destinationAssessmentSpIdsThatHaveNotBeganAssessment = destinationAssessments.Where(x => x.AssessmentToolWorkflowInstances == null || x.AssessmentToolWorkflowInstances.Count == 0)
-                        .Select(x => x.SpId).ToList();
-                    
-                    var assessmentsToBeRemoved = _syncCommandHandlerHelper.AssessmentsToBeRemoved(sourceAssessmentSpIds, destinationAssessmentSpIdsThatHaveNotBeganAssessment, destinationAssessments);
-                    await _assessmentRepository.RemoveAssesments(assessmentsToBeRemoved);
+                        var destinationAssessments = await _assessmentRepository.GetAssessments();
+                        var destinationAssessmentSpIdsThatHaveBeganAssessment = destinationAssessments.Where(x => x.AssessmentToolWorkflowInstances != null && x.AssessmentToolWorkflowInstances.Any())
+                            .Select(x => x.SpId).ToList();
+                        var destinationAssessmentSpIdsThatHaveNotBeganAssessment = destinationAssessments.Where(x => x.AssessmentToolWorkflowInstances == null || x.AssessmentToolWorkflowInstances.Count == 0)
+                            .Select(x => x.SpId).ToList();
 
-                    var validDestinationAssessmentSpIdsThatHaveNotBeganAssessment = destinationAssessmentSpIdsThatHaveNotBeganAssessment.Except(assessmentsToBeRemoved.Select(x => x.SpId)).ToList();
-                    var destinationAssessmentSpIds = destinationAssessmentSpIdsThatHaveBeganAssessment.Concat(validDestinationAssessmentSpIdsThatHaveNotBeganAssessment).ToList();
-                    var assessmentsToBeAdded = _syncCommandHandlerHelper.AssessmentsToBeAdded(sourceAssessmentSpIds, destinationAssessmentSpIds, data);
-                    await _assessmentRepository.CreateAssessments(assessmentsToBeAdded);
+                        var assessmentsToBeRemoved = _syncCommandHandlerHelper.AssessmentsToBeRemoved(sourceAssessmentSpIds, destinationAssessmentSpIdsThatHaveNotBeganAssessment, destinationAssessments);
+                        await _assessmentRepository.RemoveAssesments(assessmentsToBeRemoved);
 
-                    var existingAssessments = destinationAssessmentSpIds.Intersect(sourceAssessmentSpIds).ToList();
-                    var updatedAssessmentCount = _syncCommandHandlerHelper.UpdateAssessments(destinationAssessments, existingAssessments, data);
-                    await _assessmentRepository.SaveChanges();
-                    syncModel.NewAssessmentCount = assessmentsToBeAdded.Count;
-                    syncModel.UpdatedAssessmentCount = updatedAssessmentCount;
-                    syncModel.Synced = true;
+                        var validDestinationAssessmentSpIdsThatHaveNotBeganAssessment = destinationAssessmentSpIdsThatHaveNotBeganAssessment.Except(assessmentsToBeRemoved.Select(x => x.SpId)).ToList();
+                        var destinationAssessmentSpIds = destinationAssessmentSpIdsThatHaveBeganAssessment.Concat(validDestinationAssessmentSpIdsThatHaveNotBeganAssessment).ToList();
+                        var assessmentsToBeAdded = _syncCommandHandlerHelper.AssessmentsToBeAdded(sourceAssessmentSpIds, destinationAssessmentSpIds, data);
+                        await _assessmentRepository.CreateAssessments(assessmentsToBeAdded);
+
+                        var existingAssessments = destinationAssessmentSpIds.Intersect(sourceAssessmentSpIds).ToList();
+                        var updatedAssessmentCount = _syncCommandHandlerHelper.UpdateAssessments(destinationAssessments, existingAssessments, data);
+                        await _assessmentRepository.SaveChanges();
+                        syncModel.NewAssessmentCount = assessmentsToBeAdded.Count;
+                        syncModel.UpdatedAssessmentCount = updatedAssessmentCount;
+                        syncModel.Synced = true;
+                    }
+                    else
+                    {
+                        _logger.LogError("Single Pipeline Response data returned null");
+                        throw new ApplicationException("Single Pipeline Response data returned null");
+                    }
                 }
-                else
+                catch (Exception e)
                 {
-                    _logger.LogError("Single Pipeline Response data returned null");
-                    throw new ApplicationException("Single Pipeline Response data returned null");
+                    _logger.LogError(e, e.Message);
+                    throw new ApplicationException("Single Pipeline Data failed to sync");
+                }
+                finally
+                {
+                    await _distributedLockRepository.ReleaseLockAsync("bg_task");
                 }
             }
-            catch (Exception e)
+            else
             {
-                _logger.LogError(e,e.Message);
-                throw new ApplicationException("Single Pipeline Data failed to sync");
+                _logger.LogError("You cannot acquire the lock");
             }
+
             return syncModel;
         }
     }
